paralellize updating entities in migration, add grace period to load balancer health checks
